backend/requirements.txt
ftfy
regex
tqdm
omegaconf
pytorch-lightning
kornia
imageio-ffmpeg
einops
torch-optimizer
easydict
braceexpand
svgwrite
svgpathtools
cssutils
numba
torch-tools
visdom
pyngrok
nest_asyncio
uvicorn
